refactor(core): modularize utilities and improve code structure
- Extracted and modularized utility functions into dedicated modules for database, endpoint, URL, command, data, path, time, misc, notification, subdomain, external, and parser operations.
- Refactored imports across the codebase to use the new utility modules, replacing previous monolithic utility and common function imports.
- Improved handling of mutable default arguments in task functions for better reliability.
- Enhanced code readability and maintainability by simplifying exception handling and using more concise expressions.
- Removed unused or redundant code and streamlined logic in serializers and views.
- No changes to core business logic or external APIs; changes are focused on internal structure and maintainability.
