+++ web/api/serializers.py
@@ -1,15 +1,43 @@
 from collections import defaultdict
-from dashboard.models import *
 from django.contrib.humanize.templatetags.humanize import (naturalday, naturaltime)
 from django.db.models import F, JSONField, Value
-from recon_note.models import *
-from reNgine.common_func import *
+from reNgine.utilities.subdomain import get_interesting_subdomains
 from rest_framework import serializers
-from scanEngine.models import *
-from startScan.models import *
-from targetApp.models import *
-from dashboard.models import *
+from dashboard.models import (
+	Project,
+	SearchHistory,
+)
+from recon_note.models import (
+	TodoNote,
+)
+from scanEngine.models import (
+	EngineType,
+)
+from startScan.models import (
+	Command,
+	DirectoryFile,
+	DirectoryScan,
+	Dork,
+	Email,
+	Employee,
+	EndPoint,
+	IpAddress,
+	MetaFinderDocument,
+	Port,
+	ScanHistory,
+	Subdomain,
+	SubScan,
+	Technology,
+	Vulnerability,
+	Waf
+)
+from targetApp.models import (
+	Domain,
+	Organization,
+)
 import yaml
+
+from reNgine.definitions import ENGINE_NAMES
 
 
 class SearchHistorySerializer(serializers.ModelSerializer):
@@ -35,7 +63,7 @@ class DomainSerializer(serializers.ModelSerializer):
 	def get_vuln_count(self, obj):
 		try:
 			return obj.vuln_count
-		except:
+		except Exception:
 			return None
 
 	def get_organization(self, obj):
@@ -90,9 +118,7 @@ class SubScanResultSerializer(serializers.ModelSerializer):
 		return subscan.type
 
 	def get_engine_name(self, subscan):
-		if subscan.engine:
-			return subscan.engine.engine_name
-		return ''
+		return subscan.engine.engine_name if subscan.engine else ''
 
 
 class ReconNoteSerializer(serializers.ModelSerializer):
@@ -149,9 +175,7 @@ class SubScanSerializer(serializers.ModelSerializer):
 		return subscan.get_completed_ago()
 
 	def get_engine_name(self, subscan):
-		if subscan.engine:
-			return subscan.engine.engine_name
-		return ''
+		return subscan.engine.engine_name if subscan.engine else ''
 
 
 class CommandSerializer(serializers.ModelSerializer):
@@ -302,9 +326,7 @@ class VisualisePortSerializer(serializers.ModelSerializer):
         return f"{port.number}/{port.service_name}/{port.service_name}"
 
     def get_title(self, port):
-        if port.is_uncommon:
-            return "Uncommon Port"
-        return "Port"
+        return "Uncommon Port" if port.is_uncommon else "Port"
 
     def get_is_uncommon(self, port):
         return port.is_uncommon
@@ -409,80 +431,76 @@ class VisualiseSubdomainSerializer(serializers.ModelSerializer):
 			})
 
 		if vulnerability:
-			vulnerability_data = []
-			critical = vulnerability.filter(severity=4)
-			if critical:
-				critical_serializer = VisualiseVulnerabilitySerializer(
-					critical,
-					many=True
-				)
-				vulnerability_data.append({
-					'description': 'Critical',
-					'children': critical_serializer.data
-				})
-			high = vulnerability.filter(severity=3)
-			if high:
-				high_serializer = VisualiseVulnerabilitySerializer(
-					high,
-					many=True
-				)
-				vulnerability_data.append({
-					'description': 'High',
-					'children': high_serializer.data
-				})
-			medium = vulnerability.filter(severity=2)
-			if medium:
-				medium_serializer = VisualiseVulnerabilitySerializer(
-					medium,
-					many=True
-				)
-				vulnerability_data.append({
-					'description': 'Medium',
-					'children': medium_serializer.data
-				})
-			low = vulnerability.filter(severity=1)
-			if low:
-				low_serializer = VisualiseVulnerabilitySerializer(
-					low,
-					many=True
-				)
-				vulnerability_data.append({
-					'description': 'Low',
-					'children': low_serializer.data
-				})
-			info = vulnerability.filter(severity=0)
-			if info:
-				info_serializer = VisualiseVulnerabilitySerializer(
-					info,
-					many=True
-				)
-				vulnerability_data.append({
-					'description': 'Informational',
-					'children': info_serializer.data
-				})
-			uknown = vulnerability.filter(severity=-1)
-			if uknown:
-				uknown_serializer = VisualiseVulnerabilitySerializer(
-					uknown,
-					many=True
-				)
-				vulnerability_data.append({
-					'description': 'Unknown',
-					'children': uknown_serializer.data
-				})
-
-			if vulnerability_data:
-				return_data.append({
-					'description': 'Vulnerabilities',
-					'children': vulnerability_data
-				})
-
+			self._group_vulnerabilities_by_severity_(vulnerability, return_data)
 		if subdomain_name.screenshot_path:
 			return_data.append({
 				'description': 'Screenshot',
 				'screenshot_path': subdomain_name.screenshot_path
 			})
 		return return_data
+
+	def _group_vulnerabilities_by_severity_(self, vulnerability, return_data):
+		vulnerability_data = []
+		if critical := vulnerability.filter(severity=4):
+			critical_serializer = VisualiseVulnerabilitySerializer(
+				critical,
+				many=True
+			)
+			vulnerability_data.append({
+				'description': 'Critical',
+				'children': critical_serializer.data
+			})
+		if high := vulnerability.filter(severity=3):
+			high_serializer = VisualiseVulnerabilitySerializer(
+				high,
+				many=True
+			)
+			vulnerability_data.append({
+				'description': 'High',
+				'children': high_serializer.data
+			})
+		if medium := vulnerability.filter(severity=2):
+			medium_serializer = VisualiseVulnerabilitySerializer(
+				medium,
+				many=True
+			)
+			vulnerability_data.append({
+				'description': 'Medium',
+				'children': medium_serializer.data
+			})
+		if low := vulnerability.filter(severity=1):
+			low_serializer = VisualiseVulnerabilitySerializer(
+				low,
+				many=True
+			)
+			vulnerability_data.append({
+				'description': 'Low',
+				'children': low_serializer.data
+			})
+		if info := vulnerability.filter(severity=0):
+			info_serializer = VisualiseVulnerabilitySerializer(
+				info,
+				many=True
+			)
+			vulnerability_data.append({
+				'description': 'Informational',
+				'children': info_serializer.data
+			})
+		if uknown := vulnerability.filter(severity=-1):
+			uknown_serializer = VisualiseVulnerabilitySerializer(
+				uknown,
+				many=True
+			)
+			vulnerability_data.append({
+				'description': 'Unknown',
+				'children': uknown_serializer.data
+			})
+
+		if vulnerability_data:
+			return_data.append({
+				'description': 'Vulnerabilities',
+				'children': vulnerability_data
+			})
 
 
 class VisualiseEmailSerializer(serializers.ModelSerializer):
@@ -496,10 +514,9 @@ class VisualiseEmailSerializer(serializers.ModelSerializer):
 			'password',
 			'title'
 		]
-
 	def get_description(self, email):
 		if email.password:
-			return email.address + " > " + email.password
+			return f"{email.address} > {email.password}"
 		return email.address
 
 	def get_title(self, email):
@@ -540,10 +557,9 @@ class VisualiseEmployeeSerializer(serializers.ModelSerializer):
 		fields = [
 			'description'
 		]
-
 	def get_description(self, employee):
 		if employee.designation:
-			return employee.name + '--' + employee.designation
+			return f'{employee.name}--{employee.designation}'
 		return employee.name
 
 
@@ -563,7 +579,6 @@ class VisualiseDataSerializer(serializers.ModelSerializer):
 
 	def get_description(self, scan_history):
 		return scan_history.domain.name
-
 	def get_children(self, history):
 		scan_history = ScanHistory.objects.filter(id=history.id)
 
@@ -610,63 +625,58 @@ class VisualiseDataSerializer(serializers.ModelSerializer):
 				'children': processed_dorks})
 
 		if metainfo:
-			metainfo_data = []
-			usernames = (
-				metainfo
-				.annotate(description=F('author'))
-				.values('description')
-				.distinct()
-				.annotate(children=Value([], output_field=JSONField()))
-				.filter(author__isnull=False)
-			)
-
-			if usernames:
-				metainfo_data.append({
-					'description': 'Usernames',
-					'children': usernames})
-
-			software = (
-				metainfo
-				.annotate(description=F('producer'))
-				.values('description')
-				.distinct()
-				.annotate(children=Value([], output_field=JSONField()))
-				.filter(producer__isnull=False)
-			)
-
-			if software:
-				metainfo_data.append({
-					'description': 'Software',
-					'children': software})
-
-			os = (
-				metainfo
-				.annotate(description=F('os'))
-				.values('description')
-				.distinct()
-				.annotate(children=Value([], output_field=JSONField()))
-				.filter(os__isnull=False)
-			)
-
-			if os:
-				metainfo_data.append({
-					'description': 'OS',
-					'children': os})
-
-			if metainfo:
-				osint_data.append({
-					'description':'Metainfo',
-					'children': metainfo_data})
-
-			return_data.append({
-				'description':'OSINT',
-				'children': osint_data})
-
+			self._extracted_from_get_children_(metainfo, osint_data, return_data)
 		if osint_data:
 			return_data.append({
 				'description':'OSINT',
 				'children': osint_data})
 
+		return return_data
+
+	# TODO Rename this here and in `get_children`
+	def _extracted_from_get_children_(self, metainfo, osint_data, return_data):
+		metainfo_data = []
+		if usernames := (
+			metainfo.annotate(description=F('author'))
+			.values('description')
+			.distinct()
+			.annotate(children=Value([], output_field=JSONField()))
+			.filter(author__isnull=False)
+		):
+			metainfo_data.append({
+				'description': 'Usernames',
+				'children': usernames})
+
+		if software := (
+			metainfo.annotate(description=F('producer'))
+			.values('description')
+			.distinct()
+			.annotate(children=Value([], output_field=JSONField()))
+			.filter(producer__isnull=False)
+		):
+			metainfo_data.append({
+				'description': 'Software',
+				'children': software})
+
+		if os := (
+			metainfo.annotate(description=F('os'))
+			.values('description')
+			.distinct()
+			.annotate(children=Value([], output_field=JSONField()))
+			.filter(os__isnull=False)
+		):
+			metainfo_data.append({
+				'description': 'OS',
+				'children': os})
+
+		if metainfo:
+			osint_data.append({
+				'description':'Metainfo',
+				'children': metainfo_data})
+
+		return_data.append({
+			'description':'OSINT',
+			'children': osint_data})
 		return return_data
 
 	def process_subdomains(self, subdomains):
@@ -685,8 +695,7 @@ class VisualiseDataSerializer(serializers.ModelSerializer):
 				# Reconstruct vulnerabilities structure without duplicates
 				new_vuln_structure = []
 				for severity in ['Critical', 'High', 'Medium', 'Low', 'Informational', 'Unknown']:
-					severity_vulns = [v for k, v in vuln_dict.items() if k[1] == severity]
-					if severity_vulns:
+					if severity_vulns := [v for k, v in vuln_dict.items() if k[1] == severity]:
 						new_vuln_structure.append({
 							'description': severity,
 							'children': severity_vulns
@@ -938,11 +947,10 @@ class SubdomainSerializer(serializers.ModelSerializer):
 
 	def get_todos_count(self, subdomain):
 		return len(subdomain.get_todos.filter(is_done=False))
-
 	def get_vuln_count(self, obj):
 		try:
 			return obj.vuln_count
-		except:
+		except Exception:
 			return None
 
 
@@ -981,8 +989,6 @@ class VulnerabilitySerializer(serializers.ModelSerializer):
 			return "High"
 		elif Vulnerability.severity == 4:
 			return "Critical"
-		elif Vulnerability.severity == -1:
-			return "Unknown"
 		else:
 			return "Unknown"
