+++ web/reNgine/definitions.py
@@ -50,9 +50,7 @@ NMAP_SCRIPT = 'nmap_script'
 NMAP_SCRIPT_ARGS = 'nmap_script_args'
 NAABU_PASSIVE = 'passive'
 NAABU_RATE = 'rate'
-NUCLEI_CUSTOM_TEMPLATE = 'custom_templates'
 NUCLEI_TAGS = 'tags'
-NUCLEI_TEMPLATE = 'templates'
 NUCLEI_SEVERITY = 'severities'
 NUCLEI_CONCURRENCY = 'concurrency'
 NUCLEI_TEMPLATES = 'templates'
