+++ web/reNgine/tasks/detect.py
@@ -11,7 +11,8 @@ from reNgine.celery import app
 from reNgine.celery_custom_task import RengineTask
 from reNgine.settings import RENGINE_TOOL_PATH
 from reNgine.tasks.command import run_command
-from reNgine.common_func import get_subdomain_from_url, get_http_urls, ensure_endpoints_crawled_and_execute
+from reNgine.utilities.url import get_subdomain_from_url
+from reNgine.utilities.endpoint import get_http_urls, ensure_endpoints_crawled_and_execute
 from startScan.models import Subdomain, Waf
 
 logger = get_task_logger(__name__)
