+++ web/reNgine/tasks/vulnerability.py
@@ -8,17 +8,12 @@ from django.utils import html
 
 from reNgine.celery import app
 from reNgine.celery_custom_task import RengineTask
-from reNgine.common_func import (
-    get_http_urls,
-    ensure_endpoints_crawled_and_execute,
-    get_random_proxy,
-    generate_header_param,
-    is_iterable,
-    save_endpoint,
-    get_subdomain_from_url,
-    sanitize_url,
-    save_vulnerability,
-)
+from reNgine.utilities.endpoint import get_http_urls, ensure_endpoints_crawled_and_execute
+from reNgine.utilities.proxy import get_random_proxy
+from reNgine.utilities.command import generate_header_param
+from reNgine.utilities.data import is_iterable
+from reNgine.utilities.database import save_endpoint, save_vulnerability
+from reNgine.utilities.url import get_subdomain_from_url, sanitize_url
 from reNgine.definitions import (
     BLIND_XSS_SERVER,
     CRLFUZZ,
@@ -77,11 +72,15 @@ logger = get_task_logger(__name__)
 #-------------------------#
 
 @app.task(name='vulnerability_scan', queue='cpu_queue', bind=True, base=RengineTask)
-def vulnerability_scan(self, urls=[], ctx={}, description=None):
+def vulnerability_scan(self, urls=None, ctx=None, description=None):
     """
         This function will serve as an entrypoint to vulnerability scan.
         All other vulnerability scan will be run from here including nuclei, crlfuzz, etc
     """
+    if urls is None:
+        urls = []
+    if ctx is None:
+        ctx = {}
     logger.info('Running Vulnerability Scan Queue')
     config = self.yaml_configuration.get(VULNERABILITY_SCAN) or {}
     should_run_nuclei = config.get(RUN_NUCLEI, True)
@@ -135,11 +134,13 @@ def vulnerability_scan(self, urls=[], ctx={}, description=None):
 
 
 @app.task(name='nuclei_individual_severity_module', queue='io_queue', base=RengineTask, bind=True)
-def nuclei_individual_severity_module(self, cmd, severity, should_fetch_llm_report, ctx={}, description=None):
+def nuclei_individual_severity_module(self, cmd, severity, should_fetch_llm_report, ctx=None, description=None):
     '''
         This celery task will run vulnerability scan in parallel.
         All severities supplied should run in parallel as grouped tasks.
     '''
+    if ctx is None:
+        ctx = {}
     results = []
     logger.info(f'Running vulnerability scan with severity: {severity}')
     cmd += f' -severity {severity}'
@@ -171,7 +172,7 @@ def nuclei_individual_severity_module(self, cmd, severity, should_fetch_llm_repo
                 scan_history=self.scan,
                 target_domain=self.domain
             )
-        except:
+        except Exception:
             logger.warning(f'Subdomain {subdomain_name} was not found in the db, skipping vulnerability scan for this subdomain.')
             continue
 
@@ -215,12 +216,12 @@ def nuclei_individual_severity_module(self, cmd, severity, should_fetch_llm_repo
 
         # Send notification for all vulnerabilities except info
         url = vuln.http_url or vuln.subdomain
-        send_vuln = (
-            notif and
-            notif.send_vuln_notif and
-            vuln and
-            severity in ['low', 'medium', 'high', 'critical'])
-        if send_vuln:
+        if send_vuln := (
+            notif
+            and notif.send_vuln_notif
+            and vuln
+            and severity in ['low', 'medium', 'high', 'critical']
+        ):
             fields = {
                 'Severity': f'**{severity.upper()}**',
                 'URL': http_url,
@@ -248,20 +249,21 @@ def nuclei_individual_severity_module(self, cmd, severity, should_fetch_llm_repo
 
         # Send report to hackerone
         hackerone_query = Hackerone.objects.all()
-        send_report = (
-            hackerone_query.exists() and
-            severity not in ('info', 'low') and
-            vuln.target_domain.h1_team_handle
-        )
-        if send_report:
+        if send_report := (
+            hackerone_query.exists()
+            and severity not in ('info', 'low')
+            and vuln.target_domain.h1_team_handle
+        ):
             hackerone = hackerone_query.first()
-            if hackerone.send_critical and severity == 'critical':
+            if (
+                hackerone.send_critical
+                and severity == 'critical'
+                or hackerone.send_high
+                and severity == 'high'
+                or hackerone.send_medium
+                and severity == 'medium'
+            ):
                 send_hackerone_report.delay(vuln.id)
-            elif hackerone.send_high and severity == 'high':
-                send_hackerone_report.delay(vuln.id)
-            elif hackerone.send_medium and severity == 'medium':
-                send_hackerone_report.delay(vuln.id)
-
     # Write results to JSON file
     with open(self.output_path, 'w') as f:
         json.dump(results, f, indent=4)
@@ -298,13 +300,7 @@ def nuclei_individual_severity_module(self, cmd, severity, should_fetch_llm_repo
         ).exclude(
             severity=0
         )
-        # find all unique vulnerabilities based on path and title
-        # all unique vulnerability will go thru llm function and get report
-        # once report is got, it will be matched with other vulnerabilities and saved
-        unique_vulns = set()
-        for vuln in vulns:
-            unique_vulns.add((vuln.name, vuln.get_path()))
-
+        unique_vulns = {(vuln.name, vuln.get_path()) for vuln in vulns}
         unique_vulns = list(unique_vulns)
 
         with concurrent.futures.ThreadPoolExecutor(max_workers=DEFAULT_THREADS) as executor:
@@ -463,13 +459,17 @@ def nuclei_scan(self, urls=[], ctx={}, description=None):
 
 
 @app.task(name='dalfox_xss_scan', queue='io_queue', base=RengineTask, bind=True)
-def dalfox_xss_scan(self, urls=[], ctx=None, description=None):
+def dalfox_xss_scan(self, urls=None, ctx=None, description=None):
     """XSS Scan using dalfox
 
     Args:
         urls (list, optional): If passed, filter on those URLs.
         description (str, optional): Task description shown in UI.
     """
+    if urls is None:
+        urls = []
+    if ctx is None:
+        ctx = {}
     vuln_config = self.yaml_configuration.get(VULNERABILITY_SCAN) or {}
     should_fetch_llm_report = vuln_config.get(FETCH_LLM_REPORT, DEFAULT_GET_LLM_REPORT)
     dalfox_config = vuln_config.get(DALFOX) or {}
@@ -543,7 +543,7 @@ def dalfox_xss_scan(self, urls=[], ctx=None, description=None):
                 scan_history=self.scan,
                 target_domain=self.domain
             )
-        except:
+        except Exception:
             logger.warning(f'Subdomain {subdomain_name} was not found in the db, skipping dalfox scan for this subdomain.')
             continue
 
@@ -596,13 +596,17 @@ def dalfox_xss_scan(self, urls=[], ctx=None, description=None):
 
 
 @app.task(name='crlfuzz_scan', queue='io_queue', base=RengineTask, bind=True)
-def crlfuzz_scan(self, urls=[], ctx={}, description=None):
+def crlfuzz_scan(self, urls=None, ctx=None, description=None):
     """CRLF Fuzzing with CRLFuzz
 
     Args:
         urls (list, optional): If passed, filter on those URLs.
         description (str, optional): Task description shown in UI.
     """
+    if urls is None:
+        urls = []
+    if ctx is None:
+        ctx = {}
     vuln_config = self.yaml_configuration.get(VULNERABILITY_SCAN) or {}
     should_fetch_llm_report = vuln_config.get(FETCH_LLM_REPORT, DEFAULT_GET_LLM_REPORT)
     custom_header = vuln_config.get(CUSTOM_HEADER) or self.yaml_configuration.get(CUSTOM_HEADER)
@@ -669,7 +673,7 @@ def crlfuzz_scan(self, urls=[], ctx={}, description=None):
                 scan_history=self.scan,
                 target_domain=self.domain
             )
-        except:
+        except Exception:
             logger.warning(f'Subdomain {subdomain_name} was not found in the db, skipping crlfuzz scan for this subdomain.')
             continue
 
@@ -723,13 +727,15 @@ def crlfuzz_scan(self, urls=[], ctx={}, description=None):
 
 
 @app.task(name='s3scanner', queue='io_queue', base=RengineTask, bind=True)
-def s3scanner(self, ctx={}, description=None):
+def s3scanner(self, ctx=None, description=None):
     """Bucket Scanner
 
     Args:
         ctx (dict): Context
         description (str, optional): Task description shown in UI.
     """
+    if ctx is None:
+        ctx = {}
     input_path = str(Path(self.results_dir) / f'{self.scan_id}_s3_bucket_discovery.txt')
 
     subdomains = Subdomain.objects.filter(scan_history=self.scan)
