+++ web/reNgine/utilities/__init__.py
@@ -1,0 +1,2 @@
+# Utilities package for reNgine
+# This package contains various utility functions organized by theme
