+++ web/scanEngine/views.py
@@ -10,7 +10,7 @@ from django.contrib import messages
 from django.shortcuts import get_object_or_404, render
 from django.urls import reverse
 from rolepermissions.decorators import has_permission_decorator
-from reNgine.common_func import send_discord_message, send_slack_message, send_lark_message, send_telegram_message
+from reNgine.utilities.notification import send_discord_message, send_slack_message, send_lark_message, send_telegram_message
 from reNgine.tasks import run_command, run_gf_list
 
 from scanEngine.forms import AddEngineForm, UpdateEngineForm, AddWordlistForm, ExternalToolForm, InterestingLookupForm, NotificationForm, ProxyForm, HackeroneForm, ReportForm
